feat: 투자 성향 파악 게임 프론트엔드/백엔드 연결

See merge request s13-bigdata-dist-sub1/S13P21A301!75

+++ back/src/main/java/com/stock/survive/entity/Position.java
@@ -23,7 +23,7 @@ public class Position {
     private Long positionNo;
 
     @Column(name = "user_no", nullable = false)
-    private Integer userNo;
+    private Long userNo;
 
     @Column(name = "stock_no", nullable = false)
     private Integer stockNo;
