+++ back/src/test/java/com/stock/survive/StockInfosTest.java
@@ -53,7 +53,7 @@ public class StockInfosTest {
     @Test
     @Transactional
     public void getLatestEndPriceTest() {
-        Integer itemNo = 1; // 실제 DB에 있는 itemNo로 바꿔주세요
+        Long itemNo = 1L; // 실제 DB에 있는 itemNo로 바꿔주세요
         Integer latestPrice = stockInfosService.getLatestEndPrice(itemNo);
 
         Assertions.assertNotNull(latestPrice, "최신 종가가 존재해야 합니다");
